[refactor] : main package by MVC
