+++ app/src/main/AndroidManifest.xml
@@ -17,14 +17,14 @@
         android:roundIcon="@mipmap/lotto_logo_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".view.detail.LottoDetailActivity"></activity>
-        <activity android:name=".view.PersonalActivity" />
-        <activity android:name=".view.OpenSourceActivity" />
+        <activity android:name=".ui.detail.LottoDetailActivity"></activity>
+        <activity android:name=".ui.PersonalActivity" />
+        <activity android:name=".ui.OpenSourceActivity" />
         <activity
-            android:name=".view.MyListActivity"
+            android:name=".ui.MyListActivity"
             android:label="@string/title_activity_my_list"
             android:theme="@style/AppTheme.NoActionBar" />
-        <activity android:name=".view.SettingActivity" />
+        <activity android:name=".ui.SettingActivity" />
 
         <meta-data
             android:name="com.google.android.gms.ads.APPLICATION_ID"
@@ -50,8 +50,8 @@
             android:name=".receiver.AlarmReceiver"
             android:enabled="true" />
 
-        <activity android:name=".view.GetNumberActivity" />
-        <activity android:name=".view.main.MainActivity">
+        <activity android:name=".ui.GetNumberActivity" />
+        <activity android:name=".ui.main.MainView">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
