+++ app/src/main/java/com/lottolike/jaery/lotto/ui/detail/DetailViewModel.kt
@@ -1,9 +1,8 @@
-package com.lottolike.jaery.lotto.view.detail
+package com.lottolike.jaery.lotto.ui.detail
 
 import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import java.text.DecimalFormat
 
 class DetailViewModel : ViewModel() {
     private val _lottoNum = MutableLiveData<Int>(0)
