+++ app/src/main/java/com/lottolike/jaery/lotto/ui/main/MainController.java
@@ -1,0 +1,4 @@
+package com.lottolike.jaery.lotto.ui.main;
+
+public class MainController {
+}
