+++ app/src/main/java/com/lottolike/jaery/lotto/ui/main/MainModel.java
@@ -1,0 +1,4 @@
+package com.lottolike.jaery.lotto.ui.main;
+
+public class MainModel {
+}
