+++ app/src/main/res/layout/activity_my_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="#fff"
-    tools:context=".view.MyListActivity">
+    tools:context=".ui.MyListActivity">
 
     <Button
         android:layout_width="30dp"
