+++ app/src/main/res/layout/activity_setting.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:background="#ffffff"
     android:layout_height="match_parent"
-    tools:context=".view.SettingActivity">
+    tools:context=".ui.SettingActivity">
 
 
     <Button
