+++ app/src/main/res/layout/content_my_list.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context=".view.MyListActivity"
+    tools:context=".ui.MyListActivity"
     tools:showIn="@layout/activity_my_list">
 
 </androidx.constraintlayout.widget.ConstraintLayout>
